Manual changes: move posts page to app directory
