+++ examples/cms-wordpress/components/not-found.tsx
@@ -1,0 +1,8 @@
+'use client'
+import ErrorPage from 'next/error'
+
+const NotFound = () => {
+  return <ErrorPage statusCode={404} />
+}
+
+export default NotFound
