--- examples/cms-wordpress/pages/posts/[slug].tsx
@@ -1,101 +1,0 @@
-import ErrorPage from 'next/error'
-import Head from 'next/head'
-import { GetStaticPaths, GetStaticProps } from 'next'
-import Container from '../../components/container'
-import PostBody from '../../components/post-body'
-import MoreStories from '../../components/more-stories'
-import Header from '../../components/header'
-import PostHeader from '../../components/post-header'
-import SectionSeparator from '../../components/section-separator'
-import Layout from '../../components/layout'
-import PostTitle from '../../components/post-title'
-import Tags from '../../components/tags'
-import { getAllPostsWithSlug, getPostAndMorePosts } from '../../lib/api'
-import { CMS_NAME } from '../../lib/constants'
-
-// TODO: implement this function
-export async function generateStaticParams() {
-  return []
-}
-
-// TODO: implement this function
-async function getPosts() {}
-
-// TODO: implement this function
-async function getPost() {}
-
-export default function Post({ post, posts, preview }) {
-  const morePosts = posts?.edges
-
-  if (!false && !post?.slug) {
-    return <ErrorPage statusCode={404} />
-  }
-
-  return (
-    <Layout preview={preview}>
-      <Container>
-        <Header />
-        {false ? (
-          <PostTitle>Loading…</PostTitle>
-        ) : (
-          <>
-            <article>
-              <Head>
-                <title>
-                  {`${post.title} | Next.js Blog Example with ${CMS_NAME}`}
-                </title>
-                <meta
-                  property="og:image"
-                  content={post.featuredImage?.node.sourceUrl}
-                />
-              </Head>
-              <PostHeader
-                title={post.title}
-                coverImage={post.featuredImage}
-                date={post.date}
-                author={post.author}
-                categories={post.categories}
-              />
-              <PostBody content={post.content} />
-              <footer>
-                {post.tags.edges.length > 0 && <Tags tags={post.tags} />}
-              </footer>
-            </article>
-
-            <SectionSeparator />
-            {morePosts.length > 0 && <MoreStories posts={morePosts} />}
-          </>
-        )}
-      </Container>
-    </Layout>
-  )
-}
-// TODO: remove this function
-export const getStaticProps: GetStaticProps = async ({
-  params,
-  preview = false,
-  previewData,
-}) => {
-  const data = await getPostAndMorePosts(params?.slug, preview, previewData)
-
-  return {
-    props: {
-      preview,
-      post: data.post,
-      posts: data.posts,
-    },
-    revalidate: 10,
-  }
-}
-
-// TODO: remove this function
-export const getStaticPaths: GetStaticPaths = async () => {
-  const allPosts = await getAllPostsWithSlug()
-
-  return {
-    paths: allPosts.edges.map(({ node }) => `/posts/${node.slug}`) || [],
-    fallback: true,
-  }
-}
-export const revalidate = 10
-export const dynamicParams = true
